rename: QueryUtils 위치를 utils 디렉토리로 이동
