+++ module-core/src/main/java/com/trybe/modulecore/utils/QueryUtils.java
@@ -1,4 +1,4 @@
-package com.trybe.modulecore.post.repository;
+package com.trybe.modulecore.utils;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
